Make the Message staging model configurable
