+++ telstar/config.py
@@ -1,0 +1,5 @@
+from .com.pw import StagedMessage
+
+
+class staging:
+    model = StagedMessage
